[GITHUB-11] Adds ability to set rlimit
Adds rlimit setting and associated pam updates that will set
the max open file limit for FPM and the user that runs FPM.

Updated init script to ensure that the FPM user runs the process,
this means we only need to update the max open file limit for the
FPM user, we do not to update the root users limits at all.

Also removed unused provision tag, not sure why that was there!

Bumped version since we've got a fair few changes here.